moved dashboard components back to dashboard
